% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiChemblBioassayAllCols}
\alias{.dtiChemblBioassayAllCols}
\title{Documented column list for \code{listBioassayFields("chembl")}}
\usage{
.dtiChemblBioassayAllCols
}
\description{
Documented column list for \code{listBioassayFields("chembl")}
}
\keyword{internal}
