% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiChunk}
\alias{.dtiChunk}
\title{Split a vector into chunks of at most `size` elements}
\usage{
.dtiChunk(x, size)
}
\description{
Split a vector into chunks of at most `size` elements
}
\keyword{internal}
