% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugTargetMeta.R
\name{.dtiCombineSourceLabel}
\alias{.dtiCombineSourceLabel}
\title{Display name per source, used as a constant \code{source} column value}
\usage{
.dtiCombineSourceLabel
}
\description{
Display name per source, used as a constant \code{source} column value
}
\keyword{internal}
