% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugTargetMeta.R
\name{.dtiCompoundIdTypeMap}
\alias{.dtiCompoundIdTypeMap}
\title{Canonical compound idType (excluding \code{"name"}) -> UniChem source name}
\usage{
.dtiCompoundIdTypeMap
}
\description{
Canonical compound idType (excluding \code{"name"}) -> UniChem source name
}
\keyword{internal}
