% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiEmptyDgidb}
\alias{.dtiEmptyDgidb}
\title{Empty DGIdb data.frame with the canonical columns}
\usage{
.dtiEmptyDgidb()
}
\description{
Empty DGIdb data.frame with the canonical columns
}
\keyword{internal}
