% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiEmptyDrugs}
\alias{.dtiEmptyDrugs}
\title{Empty drug data.frame with the canonical columns}
\usage{
.dtiEmptyDrugs()
}
\description{
Empty drug data.frame with the canonical columns
}
\keyword{internal}
