% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idTranslation.R
\name{.dtiEmptyEnsemblHomology}
\alias{.dtiEmptyEnsemblHomology}
\title{Empty Ensembl homology data.frame with the canonical columns}
\usage{
.dtiEmptyEnsemblHomology()
}
\description{
Empty Ensembl homology data.frame with the canonical columns
}
\keyword{internal}
