% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiEmptyTargets}
\alias{.dtiEmptyTargets}
\title{Empty target data.frame with the canonical columns}
\usage{
.dtiEmptyTargets()
}
\description{
Empty target data.frame with the canonical columns
}
\keyword{internal}
