% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idTranslation.R
\name{.dtiEnsemblHomologyCols}
\alias{.dtiEnsemblHomologyCols}
\title{Column order for tidy Ensembl homology rows}
\usage{
.dtiEnsemblHomologyCols
}
\description{
Column order for tidy Ensembl homology rows
}
\keyword{internal}
