% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugTargetMeta.R
\name{.dtiMetaQueryBy}
\alias{.dtiMetaQueryBy}
\title{Build the native \code{queryBy} list a given source's function expects}
\usage{
.dtiMetaQueryBy(src, isGene, ids)
}
\description{
Build the native \code{queryBy} list a given source's function expects
}
\keyword{internal}
