% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiPubchemAssaySummary}
\alias{.dtiPubchemAssaySummary}
\title{Raw compound-centric assay-summary table for one or more CIDs (batched)}
\usage{
.dtiPubchemAssaySummary(cids, chunkSize = 50L)
}
\description{
Raw compound-centric assay-summary table for one or more CIDs (batched)
}
\keyword{internal}
