% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiPubchemCidProps}
\alias{.dtiPubchemCidProps}
\title{CID -> compound Title (drug/common name) + SMILES, batched}
\usage{
.dtiPubchemCidProps(cids, chunkSize = 100L)
}
\description{
CID -> compound Title (drug/common name) + SMILES, batched
}
\keyword{internal}
