% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiPubchemDrugsCols}
\alias{.dtiPubchemDrugsCols}
\title{Curated column vector for \code{fields = "core"} on \code{getPubchemDrugs()}}
\usage{
.dtiPubchemDrugsCols
}
\description{
Curated column vector for \code{fields = "core"} on \code{getPubchemDrugs()}
}
\keyword{internal}
