% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiPubchemGeneConcise}
\alias{.dtiPubchemGeneConcise}
\title{Raw concise bioactivity table for one GeneID}
\usage{
.dtiPubchemGeneConcise(geneid)
}
\description{
Raw concise bioactivity table for one GeneID
}
\keyword{internal}
