% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiPubchemGeneInfo}
\alias{.dtiPubchemGeneInfo}
\title{GeneID -> (symbol, taxid) lookup via E-utilities esummary, batched}
\usage{
.dtiPubchemGeneInfo(geneids)
}
\description{
GeneID -> (symbol, taxid) lookup via E-utilities esummary, batched
}
\keyword{internal}
