% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiPubchemTargetsCols}
\alias{.dtiPubchemTargetsCols}
\title{Curated column vector for \code{fields = "core"} on \code{getPubchemTargets()}}
\usage{
.dtiPubchemTargetsCols
}
\description{
Curated column vector for \code{fields = "core"} on \code{getPubchemTargets()}
}
\keyword{internal}
