% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idTranslation.R
\name{.dtiUniprotParseResults}
\alias{.dtiUniprotParseResults}
\title{Tidy a UniProt ID mapping job's raw results into a data.frame}
\usage{
.dtiUniprotParseResults(results)
}
\description{
Tidy a UniProt ID mapping job's raw results into a data.frame
}
\keyword{internal}
