% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idTranslation.R
\name{.dtiUniprotSubmitJob}
\alias{.dtiUniprotSubmitJob}
\title{Submit a UniProt ID mapping job}
\usage{
.dtiUniprotSubmitJob(ids, from, to, taxId = NULL)
}
\description{
Submit a UniProt ID mapping job
}
\keyword{internal}
