% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoPdbAccess.R
\name{.gtoPdbEndpoints}
\alias{.gtoPdbEndpoints}
\title{Endpoint registry for GtoPdb's bulk REST endpoint and HGNC mapping file}
\usage{
.gtoPdbEndpoints()
}
\description{
Endpoint registry for GtoPdb's bulk REST endpoint and HGNC mapping file
}
\keyword{internal}
