% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttdAccess.R
\name{.ttdFieldLookup}
\alias{.ttdFieldLookup}
\title{First-value-wins lookup for a single field of a long table}
\usage{
.ttdFieldLookup(long, field)
}
\description{
First-value-wins lookup for a single field of a long table
}
\keyword{internal}
