% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttdAccess.R
\name{.ttdParseDrugSmiles}
\alias{.ttdParseDrugSmiles}
\title{Parse P1-02's DRUGSMIL (canonical SMILES) field into a lookup}
\usage{
.ttdParseDrugSmiles(path)
}
\description{
Parse P1-02's DRUGSMIL (canonical SMILES) field into a lookup
}
\keyword{internal}
