% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttdAccess.R
\name{.ttdParseTargets}
\alias{.ttdParseTargets}
\title{Parse P1-01 into one row per TargetID (GeneName, Uniprot, TargetType)}
\usage{
.ttdParseTargets(path)
}
\description{
Parse P1-01 into one row per TargetID (GeneName, Uniprot, TargetType)
}
\keyword{internal}
