% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttdAccess.R
\name{.ttdReadLong}
\alias{.ttdReadLong}
\title{Long (id, field, value) table for one TTD flat file}
\usage{
.ttdReadLong(path, idPrefix)
}
\description{
Long (id, field, value) table for one TTD flat file
}
\keyword{internal}
