% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unichemAccess.R
\name{.unichemEndpoints}
\alias{.unichemEndpoints}
\title{Endpoint registry for UniChem's bulk table dumps}
\usage{
.unichemEndpoints()
}
\description{
Endpoint registry for UniChem's bulk table dumps
}
\keyword{internal}
