\name{drugTargetAnnot}
\alias{drugTargetAnnot}
\alias{queryBy}
\title{
drugTargetAnnot
}
\description{
 Function to query known drug-target annotations.
}
\usage{
drugTargetAnnot(queryBy=list(molType=NULL, idType=NULL, ids=NULL), cmpid_file=file.path(config$resultsPath,"cmp_ids.rds"), config=genConfig())
}
\arguments{
  \item{queryBy}{
	  A list defining the query, as described in  \code{\link{queryBy}}.
}
  \item{cmpid_file}{
	  Path to a compound ID mapping file, generated by \code{\link{cmpIdMapping}}.
	
}
	\item{config}{ General configuration. See \code{\link{genConfig}}.  }
}
\value{
	Returns the query results as a data frame.
}
\author{
	Thomas Girke
}


\seealso{
	\code{\link{queryBy}}
	\code{\link{cmpIdMapping}}
}
\examples{
	\dontrun{
		 chembldb <- "chembl_24.db" # path to full ChEMBL SQLite database
		 cmp_ids.rds <- "cmp_ids.rds"
		 cmpIdMapping(cmp_ids.rds,rerun=TRUE)
		 queryBy <- list(molType="cmp", idType="chembl_id", ids=c("CHEMBL1233058", "CHEMBL1200916", "CHEMBL437765"))
		 qresult <- drugTargetAnnot(dbpath=chembldb, queryBy, cmpid_file=cmp_ids.rds)
	}	
}
