\name{drugTargetInteractions-package}
\alias{drugTargetInteractions-package}
\alias{drugTargetInteractions}
\docType{package}
\title{
\packageTitle{drugTargetInteractions}
}
\description{
The drugTargetInteractions package provides utilities for identifying
drug-target interactions for sets of small molecule or gene/protein identifiers. 
The required drug-target interaction information is obained
from a local SQLite instance of the ChEMBL database.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{drugTargetInteractions}
\packageIndices{drugTargetInteractions}
}
\author{
\packageAuthor{drugTargetInteractions}

Maintainer: \packageMaintainer{drugTargetInteractions}
}
\keyword{ package }
