% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAccess.R
\name{.dtiDgidbSelection}
\alias{.dtiDgidbSelection}
\title{The reusable GraphQL selection for one Interaction node (no outer braces)}
\usage{
.dtiDgidbSelection()
}
\description{
The reusable GraphQL selection for one Interaction node (no outer braces)
}
\keyword{internal}
